created read-only flag and VALID_FLAGS class attributes

+++ sqlitedict.py
@@ -106,6 +106,8 @@ def decode(obj):
 
 
 class SqliteDict(DictClass):
+    VALID_FLAGS = ['c', 'r', 'w', 'n']
+
     def __init__(self, filename=None, tablename='unnamed', flag='c',
                  autocommit=False, journal_mode="DELETE"):
         """
@@ -133,6 +135,11 @@ class SqliteDict(DictClass):
         if self.in_temp:
             randpart = hex(random.randint(0, 0xffffff))[2:]
             filename = os.path.join(tempfile.gettempdir(), 'sqldict' + randpart)
+
+        if flag not in SqliteDict.VALID_FLAGS:
+            raise RuntimeError("Unrecognized flag: '{}'".format(flag))
+        self.flag = flag
+
         if flag == 'n':
             if os.path.exists(filename):
                 os.remove(filename)
@@ -218,6 +225,9 @@ class SqliteDict(DictClass):
         return decode(item[0])
 
     def __setitem__(self, key, value):
+        if self.flag == 'r':
+            raise RuntimeError('Refusing to write to read-only SqliteDict')
+
         ADD_ITEM = 'REPLACE INTO %s (key, value) VALUES (?,?)' % self.tablename
         self.conn.execute(ADD_ITEM, (key, encode(value)))
